pyproject.toml
[project]
name = "mcp-context-loading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.17.2",
    "fastapi>=0.127.0",
    "pydantic-ai>=0.0.14",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
